api/parallel_ingestion_script/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-functions-durable
python-dotenv==1.1.1
aiohttp==3.12.15
azure-kusto-data==5.0.5
azure-kusto-ingest==5.0.5